Refactor: Transition server to module-based architecture
- Grouped models, routes, and services into feature-specific modules under server/src/modules/.
- Updated all import paths across the application and tests to reflect the new structure.
- Created 'core' and 'dashboard' modules for shared/cross-cutting concerns.
- Removed old top-level models/, routes/, and services/ directories.
