+++ server/src/modules/email-target/EmailTarget.ts
